@@ -1,10 +1,12 @@
+import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:geoflutterfire/geoflutterfire.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/rendering.dart';
 import 'streambuilder_test.dart';
+import 'package:geoflutterfire/geoflutterfire.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -22,166 +24,163 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  GoogleMapController _mapController;
-  TextEditingController _latitudeController, _longitudeController;
+  Completer<GoogleMapController> _mapController = Completer();
+  TextEditingController _latitudeController = TextEditingController();
+  TextEditingController _longitudeController = TextEditingController();
 
   // firestore init
   final _firestore = FirebaseFirestore.instance;
-  Geoflutterfire geo;
-  Stream<List<DocumentSnapshot>> stream;
-  final radius = BehaviorSubject<double>.seeded(1.0);
-  Map<MarkerId, Marker> markers = <MarkerId, Marker>{};
+  late Geoflutterfire geo;
+  late Stream<List<DocumentSnapshot>> stream;
+  var radius = BehaviorSubject.seeded(100.0);
+
+  //Map markers
+  List<Marker> markers = [];
 
   @override
   void initState() {
     super.initState();
-    _latitudeController = TextEditingController();
-    _longitudeController = TextEditingController();
-
     geo = Geoflutterfire();
     GeoFirePoint center = geo.point(latitude: 12.960632, longitude: 77.641603);
     stream = radius.switchMap((rad) {
       var collectionReference = _firestore.collection('locations');
-//          .where('name', isEqualTo: 'darshan');
-      return geo.collection(collectionRef: collectionReference).within(
-          center: center, radius: rad, field: 'position', strictMode: true);
+      //          .where('name', isEqualTo: 'darshan');
+      return geo.collection(collectionRef: collectionReference).within(center: center, radius: rad, field: 'position');
 
-      /*
-      ****Example to specify nested object****
-
-      var collectionReference = _firestore.collection('nestedLocations');
-//          .where('name', isEqualTo: 'darshan');
-      return geo.collection(collectionRef: collectionReference).within(
+      /** Example to specify nested object
+          var collectionReference = _firestore.collection('nestedLocations');
+          //          .where('name', isEqualTo: 'darshan');
+          return geo.collection(collectionRef: collectionReference).within(
           center: center, radius: rad, field: 'address.location.position');
-
-      */
+       */
     });
   }
 
   @override
   void dispose() {
-    _latitudeController.dispose();
-    _longitudeController.dispose();
-    radius.close();
     super.dispose();
+    radius.close();
   }
 
   @override
   Widget build(BuildContext context) {
-    final mediaQuery = MediaQuery.of(context);
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
           title: const Text('GeoFlutterFire'),
           actions: <Widget>[
             IconButton(
-              onPressed: _mapController == null
-                  ? null
-                  : () {
-                      _showHome();
-                    },
+              onPressed: () => _showHome(),
               icon: Icon(Icons.home),
             )
           ],
         ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            Navigator.of(context).push(MaterialPageRoute(builder: (context) {
-              return StreamTestWidget();
-            }));
-          },
-          child: Icon(Icons.navigate_next),
-        ),
-        body: Container(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Center(
-                child: Card(
-                  elevation: 4,
-                  margin: EdgeInsets.symmetric(vertical: 8),
-                  child: SizedBox(
-                    width: mediaQuery.size.width - 30,
-                    height: mediaQuery.size.height * (1 / 3),
-                    child: GoogleMap(
-                      onMapCreated: _onMapCreated,
-                      initialCameraPosition: const CameraPosition(
-                        target: LatLng(12.960632, 77.641603),
-                        zoom: 15.0,
+        body: SingleChildScrollView(
+          child: Container(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                Center(
+                  child: Card(
+                    elevation: 4,
+                    margin: EdgeInsets.symmetric(vertical: 8),
+                    child: SizedBox(
+                      width: MediaQuery.of(context).size.width - 30,
+                      height: MediaQuery.of(context).size.height * (1 / 3),
+                      child: GoogleMap(
+                        onMapCreated: _onMapCreated,
+                        initialCameraPosition: const CameraPosition(
+                          target: LatLng(12.960632, 77.641603),
+                          zoom: 15.0,
+                        ),
+                        markers: markers.toSet(),
                       ),
-                      markers: Set<Marker>.of(markers.values),
                     ),
                   ),
                 ),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(top: 8.0),
-                child: Slider(
-                  min: 1,
-                  max: 200,
-                  divisions: 4,
-                  value: _value,
-                  label: _label,
-                  activeColor: Colors.blue,
-                  inactiveColor: Colors.blue.withOpacity(0.2),
-                  onChanged: (double value) => changed(value),
-                ),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: <Widget>[
-                  Container(
-                    width: 100,
-                    child: TextField(
-                      controller: _latitudeController,
-                      keyboardType: TextInputType.number,
-                      textInputAction: TextInputAction.next,
-                      decoration: InputDecoration(
-                          labelText: 'lat',
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                          )),
-                    ),
+                Padding(
+                  padding: const EdgeInsets.only(top: 8.0),
+                  child: Slider(
+                    min: 1,
+                    max: 200,
+                    divisions: 4,
+                    value: _value,
+                    label: _label,
+                    activeColor: Colors.blue,
+                    inactiveColor: Colors.blue.withOpacity(0.2),
+                    onChanged: (double value) => changed(value),
                   ),
-                  Container(
-                    width: 100,
-                    child: TextField(
-                      controller: _longitudeController,
-                      keyboardType: TextInputType.number,
-                      decoration: InputDecoration(
-                          labelText: 'lng',
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                          )),
-                    ),
-                  ),
-                  MaterialButton(
-                    color: Colors.blue,
-                    onPressed: () {
-                      final lat = double.parse(_latitudeController.text);
-                      final lng = double.parse(_longitudeController.text);
-                      _addPoint(lat, lng);
-                    },
-                    child: const Text(
-                      'ADD',
-                      style: TextStyle(color: Colors.white),
-                    ),
-                  )
-                ],
-              ),
-              MaterialButton(
-                color: Colors.amber,
-                child: const Text(
-                  'Add nested ',
-                  style: TextStyle(color: Colors.white),
                 ),
-                onPressed: () {
-                  final lat = double.parse(_latitudeController.text);
-                  final lng = double.parse(_longitudeController.text);
-                  _addNestedPoint(lat, lng);
-                },
-              )
-            ],
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: <Widget>[
+                    Container(
+                      width: 100,
+                      child: TextField(
+                        controller: _latitudeController,
+                        keyboardType: TextInputType.number,
+                        textInputAction: TextInputAction.next,
+                        decoration: InputDecoration(
+                            labelText: 'lat',
+                            border: OutlineInputBorder(
+                              borderRadius: BorderRadius.circular(8),
+                            )),
+                      ),
+                    ),
+                    Container(
+                      width: 100,
+                      child: TextField(
+                        controller: _longitudeController,
+                        keyboardType: TextInputType.number,
+                        decoration: InputDecoration(
+                            labelText: 'lng',
+                            border: OutlineInputBorder(
+                              borderRadius: BorderRadius.circular(8),
+                            )),
+                      ),
+                    ),
+                    MaterialButton(
+                      color: Colors.blue,
+                      onPressed: () {
+                        double lat = double.parse(_latitudeController.text);
+                        double lng = double.parse(_longitudeController.text);
+                        _addPoint(lat, lng);
+                      },
+                      child: Text(
+                        'ADD',
+                        style: TextStyle(color: Colors.white),
+                      ),
+                    )
+                  ],
+                ),
+                MaterialButton(
+                  color: Colors.amber,
+                  child: Text(
+                    'Add nested ',
+                    style: TextStyle(color: Colors.white),
+                  ),
+                  onPressed: () {
+                    double lat = double.parse(_latitudeController.text);
+                    double lng = double.parse(_longitudeController.text);
+                    _addNestedPoint(lat, lng);
+                  },
+                ),
+                MaterialButton(
+                  color: Colors.blueGrey,
+                  child: Text(
+                    'Stream Test ',
+                    style: TextStyle(color: Colors.white),
+                  ),
+                  onPressed: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => StreamTestWidget()),
+                    );
+                  },
+                )
+
+              ],
+            ),
           ),
         ),
       ),
@@ -189,18 +188,15 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _onMapCreated(GoogleMapController controller) {
-    setState(() {
-      _mapController = controller;
-//      _showHome();
-      //start listening after map is created
-      stream.listen((List<DocumentSnapshot> documentList) {
-        _updateMarkers(documentList);
-      });
+    _mapController.complete(controller);
+    stream.listen((List<DocumentSnapshot> documentList) {
+      _updateMarkers(documentList);
     });
   }
 
-  void _showHome() {
-    _mapController.animateCamera(CameraUpdate.newCameraPosition(
+  void _showHome() async {
+    final GoogleMapController controller = await _mapController.future;
+    controller.animateCamera(CameraUpdate.newCameraPosition(
       const CameraPosition(
         target: LatLng(12.960632, 77.641603),
         zoom: 15.0,
@@ -210,10 +206,11 @@ class _MyAppState extends State<MyApp> {
 
   void _addPoint(double lat, double lng) {
     GeoFirePoint geoFirePoint = geo.point(latitude: lat, longitude: lng);
-    _firestore
-        .collection('locations')
-        .add({'name': 'random name', 'position': geoFirePoint.data}).then((_) {
+    _firestore.collection('locations').add({'name': 'random name', 'position': geoFirePoint.data}).then((_) {
       print('added ${geoFirePoint.hash} successfully');
+    });
+    setState(() {
+
     });
   }
 
@@ -231,26 +228,24 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _addMarker(double lat, double lng) {
-    final id = MarkerId(lat.toString() + lng.toString());
-    final _marker = Marker(
-      markerId: id,
+    var _marker = Marker(
+      markerId: MarkerId(UniqueKey().toString()),
       position: LatLng(lat, lng),
       icon: BitmapDescriptor.defaultMarkerWithHue(BitmapDescriptor.hueViolet),
-      infoWindow: InfoWindow(title: 'latLng', snippet: '$lat,$lng'),
     );
     setState(() {
-      markers[id] = _marker;
+      markers.add(_marker);
     });
   }
 
   void _updateMarkers(List<DocumentSnapshot> documentList) {
     documentList.forEach((DocumentSnapshot document) {
-      final GeoPoint point = document.data()['position']['geopoint'];
+      GeoPoint point = document['position']['geopoint'];
       _addMarker(point.latitude, point.longitude);
     });
   }
 
-  double _value = 20.0;
+  double _value = 40.0;
   String _label = '';
 
   changed(value) {
@@ -258,7 +253,7 @@ class _MyAppState extends State<MyApp> {
       _value = value;
       _label = '${_value.toInt().toString()} kms';
       markers.clear();
+      radius.add(value);
     });
-    radius.add(value);
   }
 }
