@@ -9,7 +9,7 @@ class StreamTestWidget extends StatefulWidget {
 }
 
 class _StreamTestWidgetState extends State<StreamTestWidget> {
-  Stream<List<DocumentSnapshot>> stream;
+  late Stream<List<DocumentSnapshot>> stream;
   final _firestore = FirebaseFirestore.instance;
   final geo = Geoflutterfire();
 
@@ -22,8 +22,9 @@ class _StreamTestWidgetState extends State<StreamTestWidget> {
     final center = geo.point(latitude: 12.960632, longitude: 77.641603);
     stream = radius.switchMap((rad) {
       var collectionReference = _firestore.collection('locations');
-      return geo.collection(collectionRef: collectionReference).within(
-          center: center, radius: rad, field: 'position', strictMode: true);
+      return geo
+          .collection(collectionRef: collectionReference)
+          .within(center: center, radius: rad, field: 'position', strictMode: true);
     });
   }
 
@@ -35,31 +36,33 @@ class _StreamTestWidgetState extends State<StreamTestWidget> {
         children: <Widget>[
           StreamBuilder(
             stream: stream,
-            builder: (BuildContext context,
-                AsyncSnapshot<List<DocumentSnapshot>> snapshots) {
-              if (snapshots.connectionState == ConnectionState.active &&
-                  snapshots.hasData) {
+            builder: (BuildContext context, AsyncSnapshot<List<DocumentSnapshot>> snapshots) {
+              if (snapshots.connectionState == ConnectionState.active && snapshots.hasData) {
                 print('data ${snapshots.data}');
                 return Container(
                   height: MediaQuery.of(context).size.height * 2 / 3,
                   child: ListView.builder(
                     itemBuilder: (context, index) {
-                      final doc = snapshots.data[index];
-                      final data = doc.data();
-                      print(
-                          'doc with id ${doc.id} distance ${data['distance']}');
-                      GeoPoint point = data['position']['geopoint'];
-                      return ListTile(
-                        title: Text(
-                          doc.id,
-                          style: const TextStyle(fontWeight: FontWeight.bold),
-                        ),
-                        subtitle: Text('${point.latitude}, ${point.longitude}'),
-                        trailing: Text(
-                            '${data['documentType'] == DocumentChangeType.added ? 'Added' : 'Modified'}'),
-                      );
+                      DocumentSnapshot<Map<String, dynamic>>? doc =
+                      snapshots.data![index] as DocumentSnapshot<Map<String, dynamic>>;
+                      Map<String, dynamic>? data = doc.data();
+                      if (data == null) {
+                        return SizedBox();
+                      } else {
+                        print('doc with id ${doc.id} distance ${data['distance']}');
+                        GeoPoint point = data['position']['geopoint'];
+                        return ListTile(
+                          title: Text(
+                            doc.id,
+                            style: const TextStyle(fontWeight: FontWeight.bold),
+                          ),
+                          subtitle: Text('${point.latitude}, ${point.longitude}'),
+                          trailing: Text(
+                              '${data['documentType'] == DocumentChangeType.added ? 'Added' : 'Modified'}'),
+                        );
+                      }
                     },
-                    itemCount: snapshots.data.length,
+                    itemCount: snapshots.data!.length,
                   ),
                 );
               } else {
