@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class DistanceDocSnapshot {
-  final DocumentSnapshot documentSnapshot;
+  final DocumentSnapshot<Map<String, dynamic>> documentSnapshot;
   double distance;
 
   DistanceDocSnapshot(this.documentSnapshot, this.distance);
